Added Sprite for Boss
Added the sprite for the Boss. (Need to work on setting him up)

Added block timer (so you can't block forever, so it is like a parry [sort of...])

Added health pack for life regeneration.

Added Enemy Spawns (enemys show up when player hits a trigger) on Level 3 as well as Enemy Replacement (enemys show up when a specific enemy is killed)